Moved the parallelogram tests from utilities to parallelogram file
